requirements.txt
flask
requests
pandas
kafka-python